+++ clang-tools-extra/docs/ReleaseNotes.md
@@ -125,6 +125,12 @@ infrastructure are described first, followed by tool-specific sections.
   Finds instances where `py::call_guard<py::gil_scoped_release>()` is passed
   alongside `py::init(...)` in pybind11 class definitions.
 
+- New {doc}`pybind-cast-in-lock
+  <clang-tidy/checks/pybind/cast-in-lock>` check.
+
+  Finds calls to `pybind11::cast` or `pybind11::handle::cast` executed while holding
+  a synchronization lock guard.
+
 - New {doc}`readability-redundant-zero-initializer
   <clang-tidy/checks/readability/redundant-zero-initializer>` check.